% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSSP_fun.R
\name{DSSP.predict}
\alias{DSSP.predict}
\title{Function to generate samples from the posterior predictive distribution of y for new locations x.new}
\usage{
DSSP.predict(dssp.model, x.pred)
}
\arguments{
\item{dssp.model}{the model output from \code{DSSP()}.}

\item{x.pred}{a matrix of spatial coordinates to predict new values of y.}
}
\value{
A matrix with N columns containing N values from the posterior predictive distribution.
}
\description{
This function samples from the log-posterior of all parameters in the model
}
\examples{
## Use the Meuse River dataset from the package 'gstat'

library(sp)
library(gstat)
data(meuse.all)
coordinates(meuse.all) <- ~ x + y

f <- function(x) -x ## log-prior for exponential distribution for the smoothing parameter

## Draw 100 samples from the posterior of eta given the data y.
OUTPUT <- DSSP(
  formula = log(zinc) ~ 1, data = meuse.all[1:155, ], N = 100,
  pars = c(0.001, 0.001), log_prior = f
)
Y.PRED <- DSSP.predict(OUTPUT, meuse.all[156:164, ])
}
\keyword{prior,}
\keyword{spatial}
\keyword{splines}
\keyword{thin-plate}
