% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSSP_fun.R
\name{eta.post}
\alias{eta.post}
\title{Log-posterior of smoothing parameter}
\usage{
eta.post(eta, nd, ev, Q, log_prior)
}
\arguments{
\item{eta}{the value of the smoothing parameter, note that eta > 0.}

\item{nd}{the rank of the precision matrix, the default value is n-3 for spatial data.}

\item{ev}{eigenvalues of the precision matrix spatial prior from the function make.m().}

\item{Q}{the data vector from the make.Q function.}

\item{log_prior}{the log prior density of smoothing parameter.}
}
\value{
The log of the posterior density of eta evaluated at eta = x.
}
\description{
This function evaluates the log-posterior density of the smoothing parameter from the
   direct sampling spatial prior (DSSP).
}
\examples{
## Use the Meuse River dataset from the package 'gstat'

library(sp)
library(gstat)
data(meuse.all)
coordinates(meuse.all)<-~x+y
X<-scale(coordinates(meuse.all))
tmp<-make.M(X)

EV<-tmp$M.eigen$values
V<-tmp$M.eigen$vectors

Y<-scale(log(meuse.all$zinc))
Q<-make.Q(Y,V)

ND<-nrow(X)-3
f<-function(x) -x ## log of the exponential density with E(x) = 1.

## Evaluate the log posterior of eta|y at eta=0.3
## This function is mainly for internal use only.

eta.post(0.3,ND,EV,Q,f)
}
\keyword{prior,}
\keyword{spatial}
\keyword{splines}
\keyword{thin-plate}
