% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSSP_fun.R
\name{make.Q}
\alias{make.Q}
\title{Data Vector Function}
\usage{
make.Q(y, V)
}
\arguments{
\item{y}{a vector of observed data.}

\item{V}{a matrix of eigen vectors from the precision matrix computed by \code{make.M()}.}
}
\value{
A vector Q = y' V which is used in subsequent functions for sampling
 from the  joint posterior distribution.
}
\description{
This function creates the the data vector for evaluating the joint posterior distribution of 
the direct sampling spatial prior (DSSP) model.
}
\examples{
## Use the Meuse River dataset from the package 'gstat'

library(sp)
library(gstat)
data(meuse.all)
coordinates(meuse.all) <- ~ x + y
X <- scale(coordinates(meuse.all))
tmp <- make.M(X)

EV <- tmp$M.eigen$values
V <- tmp$M.eigen$vectors

Y <- scale(log(meuse.all$zinc))
Q <- make.Q(Y, V)
}
\keyword{prior,}
\keyword{spatial}
\keyword{splines}
\keyword{thin-plate}
