% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSSP_fun.R
\name{sample.eta}
\alias{sample.eta}
\title{Function to sample from the posterior of the smoothing parameter eta conditioned on the data
   y.}
\usage{
sample.eta(N, nd, ev, Q, log_prior, UL = 1000)
}
\arguments{
\item{N}{the number of samples desired.}

\item{nd}{the rank of the precision matrix, the default value is n-3 for spatial data.}

\item{ev}{eigenvalues of the precision matrix spatial prior from the function make.m().}

\item{Q}{the data vector from the make.Q function.}

\item{log_prior}{the log prior density of smoothing parameter.}

\item{UL}{the upper limit for the smoothing parameter value; used for the
ratio-of-uniform sampler, default is 1000.}
}
\value{
N samples drawn from the posterior of eta given the data y \eqn{\pi(eta | y)}.
}
\description{
This function samples from the log-posterior density of the smoothing parameter from the
  thin-plate splines based spatial prior using a ratio-of-uniform sampler.
}
\examples{
## Use the Meuse River dataset from the package 'gstat'

library(sp)
library(gstat)
data(meuse.all)
coordinates(meuse.all)<-~x+y
X<-scale(coordinates(meuse.all))
tmp<-make.M(X)

EV<-tmp$M.eigen$values
V<-tmp$M.eigen$vectors

Y<-scale(log(meuse.all$zinc))
Q<-make.Q(Y,V)

ND<-nrow(X)-3
f<-function(x) -x ## log-prior for exponential distribution for the smoothing parameter
## Draw 100 samples from the posterior of eta given the data y.

sample.eta(100,ND,EV,Q,f,UL=1000)
}
\keyword{prior,}
\keyword{spatial}
\keyword{splines}
\keyword{thin-plate}
