% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSSP_fun.R
\name{DSSP.predict}
\alias{DSSP.predict}
\title{Function to generate samples from the posterior predictive distribution of y for new locations x.new}
\usage{
DSSP.predict(dssp.model, x.pred, ncores = 1)
}
\arguments{
\item{dssp.model}{the model output from DSSP()}

\item{x.pred}{a matrix of spatial coordinates to predict new values of y}

\item{ncores}{number of cores to use when sampling from the posterior predictive distribution,
defaults to 1. This is an optional argument to pass to rmvn() from the mvnfast package
It requires that the user's system has OpenMP installed and packages are set to build with
OpenMp enabled.  Please see the documentation for the mvnfast package for further details.}
}
\value{
A matrix with N columns containing N values from the posterior predictive distribution.
}
\description{
This function samples from the log-posterior of all parameters in the model
}
\examples{
#'## Use the Meuse River dataset from the package 'gstat'

library(sp)
library(gstat)
data(meuse.all)
coordinates(meuse.all)<-~x+y
X<-scale(coordinates(meuse.all))

Y<-scale(log(meuse.all$zinc))

Y.train<-Y[1:155]
X.train<-X[1:155,]
X.pred<-X[156:164,]

f<-function(x) -x ## log-prior for exponential distribution for the smoothing parameter

## Draw 100 samples from the posterior of eta given the data y.

OUTPUT<-DSSP(100,X.train,Y.train,f,pars=c(0.001,0.001),fitted.values=FALSE,ncores=1)

Y.PRED<-DSSP.predict(OUTPUT,X.pred,ncores=1)
}
\keyword{prior,}
\keyword{spatial}
\keyword{splines}
\keyword{thin-plate}
