% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSSP_fun.R
\name{DSSP}
\alias{DSSP}
\title{Wrapper function to draw samples from the Direct Sampling Spatial Prior (DSSP) model}
\usage{
DSSP(N, x, y, log_prior, pars, ncores = 1)
}
\arguments{
\item{N}{is the number of random samples to be drawn from the joint posterior for eta, delta, and nu}

\item{x}{a matrix of spatial coordinates}

\item{y}{vector of obsered data}

\item{log_prior}{the log-prior density for the prior of the smoothing parameter of the
direct sampling spatial prior (DSSP).}

\item{pars}{a vector of the prior shape and rate parameters for the
inverse-gamma prior distribution of delta, the variance parameter for the
Gaussian likelihood.}

\item{ncores}{number of cores to use when sampling from the distribution of nu, defaults to 1.
This is an optional argument to pass to rmvn() from the mvnfast package, it requries that
the user's system has OpenMP installed and pacakges are set to build with OpenMp enabled.
Please see the documentation for the mvnfast package for further details.}
}
\value{
A list containing N samples of nu, eta, delta, and the original data X and Y.
}
\description{
This function samples from the log-posterior of all parameters in the model and returns a list
   object containing the samples. It performs a few compatability checks on the inputs, then
   calls the sample.eta(), sample.delta(), and sample.nu().
}
\details{
The direct sampling spatial prior model assumes that the spatial model can be written
 as the likelihood parameterised with mean vector nu and variance delta
 \deqn{(y | nu, delta) ~ N(nu, delta * I)}
 where I is the identity matrix.  The prior for the vector of spatial effects nu is
 improper but is proportional to
 \deqn{\pi(nu | eta) propto (det(M)/2\pi)^{1/2} * exp(-eta nu' M nu/2),}
 the prior for delta is assumed to be a inverse-gamma distribution
 \deqn{(delta) ~ IG(a,b)}
 and the prior for eta can be specificed for the user as any valid density function for eta > 0.
}
\examples{
## Use the Meuse River dataset from the package 'gstat'

library(sp)
library(gstat)
data(meuse.all)
coordinates(meuse.all)<-~x+y
X<-scale(coordinates(meuse.all))

Y<-scale(log(meuse.all$zinc))

f<-function(x) -x ## log-prior for exponential distribution for the smoothing parameter

## Draw 100 samples from the posterior of eta given the data y.

OUTPUT<-DSSP(100,X,Y,f,pars=c(0.001,0.001),ncores=1)
}
\keyword{prior,}
\keyword{spatial}
\keyword{splines}
\keyword{thin-plate}
