% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSSP_fun.R
\name{sample.delta}
\alias{sample.delta}
\title{Function to sample from the posterior of the variance parameter}
\usage{
sample.delta(eta, nd, ev, Q, pars)
}
\arguments{
\item{eta}{samples of the smoothing parameter from the sample.eta function.}

\item{nd}{the rank of the precision matrix, the default value is n-3 for spatial data.}

\item{ev}{eigenvalues of the precision matrix spatial prior from the function make.m().}

\item{Q}{the data vector from the make.Q function.}

\item{pars}{a vector of the prior shape and rate parameters for the
inverse-gamma prior distribution of delta.}
}
\value{
N samples drawn from the posterior of \eqn{\pi(delta | eta, y)}.
}
\description{
This function samples from the log-posterior density of the variance parameter from the likelihood
}
\examples{
## Use the Meuse River dataset from the package 'gstat'

library(sp)
library(gstat)
data(meuse.all)
coordinates(meuse.all)<-~x+y
X<-scale(coordinates(meuse.all))
tmp<-make.M(X)

EV<-tmp$M.eigen$values
V<-tmp$M.eigen$vectors

Y<-scale(log(meuse.all$zinc))
Q<-make.Q(Y,V)

ND<-nrow(X)-3
f<-function(x) -x ## log-prior for exponential distribution for the smoothing parameter
## Draw 100 samples from the posterior of eta given the data y.

ETA<-sample.eta(100,ND,EV,Q,f,UL=1000)
DELTA<-sample.delta(ETA,ND,EV,Q,pars=c(0.001,0.001))
}
\keyword{prior,}
\keyword{spatial}
\keyword{splines}
\keyword{thin-plate}
