% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSSP_fun.R
\name{sample.nu}
\alias{sample.nu}
\title{Function to sample from the posterior of the spatial effects}
\usage{
sample.nu(y, eta, delta, ev, V, ncores = 1)
}
\arguments{
\item{y}{vector of observed data}

\item{eta}{samples of the smoothing parameter from the sample.eta function}

\item{delta}{samples of the variance parameter from the sample.delta function}

\item{ev}{eigenvalues of the precision matrix spatial prior from the function make.M()}

\item{V}{eigenvectors of the precision matrix spatial prior from the function make.M()}

\item{ncores}{number of cores to use when sampling from the distribution of nu, defaults to 1.
This is an optional argument to pass to rmvn() from the mvnfast package, it requries that
the user's system has OpenMP installed and pacakges are set to build with OpenMp enabled.
Please see the documentation for the mvnfast package for further details.}
}
\value{
A matrix of samples with each column a random draw from the posterior
   of the spatial effects from the DSSP model \eqn{\pi(nu | eta, delta, y)}.
}
\description{
This function samples from the posterior density of the spatial effects from the direct sampling
 spatial prior (DSSP) model.
}
\examples{
## Use the Meuse River dataset from the package 'gstat'

library(sp)
library(gstat)
data(meuse.all)
coordinates(meuse.all)<-~x+y
X<-scale(coordinates(meuse.all))
tmp<-make.M(X)

EV<-tmp$M.eigen$values
V<-tmp$M.eigen$vectors

Y<-scale(log(meuse.all$zinc))
Q<-make.Q(Y,V)

ND<-nrow(X)-3
f<-function(x) -x ## log-prior for exponential distribution for the smoothing parameter
## Draw 100 samples from the posterior of eta given the data y.

ETA<-sample.eta(100,ND,EV,Q,f,UL=1000)
DELTA<-sample.delta(ETA,ND,EV,Q,pars=c(0.001,0.001))
NU<-sample.nu(Y,ETA,DELTA,EV,V,ncores=1)
}
\keyword{prior,}
\keyword{spatial}
\keyword{splines}
\keyword{thin-plate}
